requirements.txt
flask==2.2.3
werkzeug==2.2.3
python-docx==0.8.11
beautifulsoup4==4.10.0
pyjwt==2.3.0
python-dotenv==0.19.2
gunicorn==20.1.0 